+++ cypress/integration/core/image-overlay.spec.ts
@@ -58,6 +58,21 @@ describe("Image Overlay", () => {
     cy.get('body').realClick({ y: 50, x: 50 });
     cy.get('dso-image-overlay').shadow().find('.wrapper > img').should('not.exist');
   });
+
+  it.only('should only have overlay available if image is zoomable' , () => {
+    cy.get('dso-image-overlay')
+      .find('img')
+      .invoke('attr', 'src', 'images/houtkachel-of-open-haard-infographic.jpg')
+      .invoke('attr', 'width', 400)
+      .get('@open-button')
+      .should('exist')
+      .get('dso-image-overlay')
+      .find('img')
+      .invoke('attr', 'width', 720) // this is the image natural width
+      .wait(100)
+      .get('@open-button')
+      .should('not.exist');
+  });
 });
 
 function openOverlay() {